Swap replace original template lib files with versioning-enabled copies and update all references
