+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/components/dialogs/CreateDraftConfirmDialog/actions.ts
@@ -4,7 +4,7 @@ import { revalidatePath } from "next/cache";
 
 import { redirect } from "next/navigation";
 
-import { createDraftVersionForTemplate } from "@lib/templates/versioning/mutations/createDraftForTemplate";
+import { createDraftVersionForTemplate } from "@root/lib/templates/mutations/createDraftForTemplate";
 
 import { AuthenticatedAction } from "@lib/actions";
 import { type FormRecord } from "@lib/types";
