+++ lib/templates/mutations/publishTemplate.ts
@@ -1,26 +1,42 @@
 import { formCache } from "@lib/cache/formCache";
-import { prisma, prismaErrors } from "@gcforms/database";
+import { prisma, prismaErrors, Prisma } from "@gcforms/database";
 import { FormRecord } from "@lib/types";
 import { deleteDraftFormResponses } from "@lib/vault";
-import { getFullTemplateByID } from "../queries/getFullTemplateByID";
 import { TemplateAlreadyPublishedError } from "../internal/errors";
-import { parseTemplate } from "../internal";
+import { getTemplateVersionState } from "../queries/getTemplateVersionState";
+import { getFullTemplateByID } from "../queries/getFullTemplateByID";
+import {
+  getBuilderVersion,
+  getTemplateJsonConfigMirrorData,
+  parseTemplate,
+  templateRecordInclude,
+} from "../internal";
+import { TEMPLATE_VERSION_STATUS } from "../internal/types";
 import { UpdateIsPublishedCommand } from "../types";
-import { isTemplateVersioningEnabled } from "../versioning/internal";
-import { publishTemplate as publishTemplateVersioningEnabled } from "../versioning/mutations/publishTemplate";
 import { authorizeForCommand } from "./shared/authorizeForCommand";
 import { logTemplateUpdateEvent } from "./shared/logTemplateUpdateEvent";
 
 export async function publishTemplate(
   command: UpdateIsPublishedCommand
 ): Promise<FormRecord | null> {
-  if (await isTemplateVersioningEnabled()) {
-    return publishTemplateVersioningEnabled(command);
-  }
-
   const { user } = await authorizeForCommand(command);
 
-  if (command.isPublished && process.env.APP_ENV !== "test") {
+  const templateVersionState = await getTemplateVersionState(command.formId);
+  const supportsVersionedPublishing = Boolean(
+    (templateVersionState?.currentDraftVersionId ||
+      templateVersionState?.currentPublishedVersionId) &&
+    templateVersionState?.isPublished !== undefined
+  );
+
+  if (supportsVersionedPublishing && !command.isPublished) {
+    throw new Error("Unpublishing template versions is not supported.");
+  }
+
+  if (
+    command.isPublished &&
+    process.env.APP_ENV !== "test" &&
+    (!supportsVersionedPublishing || templateVersionState?.isPublished === false)
+  ) {
     try {
       await deleteDraftFormResponses(command.formId);
     } catch (e) {
@@ -32,33 +48,88 @@ export async function publishTemplate(
     }
   }
 
-  const updatedTemplate = await prisma.template
-    .update({
-      where: {
-        id: command.formId,
-        isPublished: {
-          not: command.isPublished,
-        },
-      },
-      data: {
-        isPublished: command.isPublished,
-        publishReason: command.publishReason,
-        publishFormType: command.publishFormType,
-        publishDesc: command.publishDescription,
-        lastEditedBy: { connect: { id: user.id } },
-      },
-      include: {
-        deliveryOption: true,
-        lastEditedBy: {
-          select: {
-            name: true,
-          },
-        },
-      },
-    })
-    .catch((e) => prismaErrors(e, null));
+  const updatedTemplate = supportsVersionedPublishing
+    ? await prisma
+        .$transaction(async (tx) => {
+          const template = await tx.template.findUnique({
+            where: {
+              id: command.formId,
+            },
+            include: templateRecordInclude,
+          });
 
-  if (updatedTemplate === null) return null;
+          if (!template) return null;
+
+          if (!template.currentDraftVersion) {
+            return template.isPublished ? template : null;
+          }
+
+          const now = new Date();
+
+          if (template.currentPublishedVersionId) {
+            await tx.templateVersion.update({
+              where: {
+                id: template.currentPublishedVersionId,
+              },
+              data: {
+                status: TEMPLATE_VERSION_STATUS.SUPERSEDED,
+                supersededAt: now,
+              },
+            });
+          }
+
+          const publishedVersion = await tx.templateVersion.update({
+            where: {
+              id: template.currentDraftVersion.id,
+            },
+            data: {
+              status: TEMPLATE_VERSION_STATUS.PUBLISHED,
+              publishedAt: now,
+              publishedByUserId: user.id,
+              publishReason: command.publishReason,
+            },
+            select: {
+              id: true,
+              jsonConfig: true,
+            },
+          });
+
+          return tx.template.update({
+            where: {
+              id: command.formId,
+            },
+            data: {
+              isPublished: true,
+              ...getTemplateJsonConfigMirrorData(publishedVersion.jsonConfig as Prisma.JsonObject),
+              currentPublishedVersionId: publishedVersion.id,
+              currentDraftVersionId: null,
+              publishReason: command.publishReason,
+              publishFormType: command.publishFormType || template.publishFormType || "",
+              publishDesc: command.publishDescription || template.publishDesc || "",
+            },
+            include: templateRecordInclude,
+          });
+        })
+        .catch((e) => prismaErrors(e, null))
+    : await prisma.template
+        .update({
+          where: {
+            id: command.formId,
+            isPublished: {
+              not: command.isPublished,
+            },
+          },
+          data: {
+            isPublished: command.isPublished,
+            publishReason: command.publishReason,
+            publishFormType: command.publishFormType,
+            publishDesc: command.publishDescription,
+          },
+          include: templateRecordInclude,
+        })
+        .catch((e) => prismaErrors(e, null));
+
+  if (updatedTemplate === null) return updatedTemplate;
 
   if (formCache.cacheAvailable) formCache.invalidate(command.formId);
 
@@ -66,7 +137,10 @@ export async function publishTemplate(
     action: command.action,
     command,
     user,
+    isRepublish: Boolean(templateVersionState?.currentPublishedVersionId),
   });
 
-  return parseTemplate(updatedTemplate);
+  return parseTemplate(updatedTemplate, {
+    version: getBuilderVersion(updatedTemplate),
+  });
 }
