+++ lib/templates/queries/getFullTemplateByID.ts
@@ -2,24 +2,13 @@ import { FormRecord } from "@lib/types";
 import { prisma, prismaErrors } from "@gcforms/database";
 import { authorization } from "@lib/privileges";
 import { AuditLogAccessDeniedDetails, logEvent } from "@lib/auditLogs";
-import { parseTemplate } from "../internal";
-import { isTemplateVersioningEnabled } from "../versioning/internal";
-import { getFullTemplateByID as getFullTemplateByIDVersioningEnabled } from "@lib/templates/versioning/queries/getFullTemplateByID";
+import { getBuilderVersion, parseTemplate, templateRecordInclude } from "../internal";
 
-/**
- * Get a form template by ID (includes full template information but requires view permission)
- * @param formID ID of form template
- * @returns FormRecord
- */
 export async function getFullTemplateByID(
   formID: string,
   allowDeleted?: boolean,
   versionNumber?: number
 ): Promise<FormRecord | null> {
-  if (await isTemplateVersioningEnabled()) {
-    return getFullTemplateByIDVersioningEnabled(formID, allowDeleted, versionNumber);
-  }
-
   try {
     const { user } = await authorization.canViewForm(formID, allowDeleted).catch((e) => {
       logEvent(
@@ -37,23 +26,37 @@ export async function getFullTemplateByID(
           id: formID,
           ttl: allowDeleted ? { not: null } : null,
         },
-        include: {
-          deliveryOption: true,
-          lastEditedBy: {
-            select: {
-              name: true,
-            },
-          },
-        },
+        include: templateRecordInclude,
       })
       .catch((e) => prismaErrors(e, null));
 
     if (!template) return null;
 
+    const requestedVersion = versionNumber
+      ? await prisma.templateVersion
+          .findFirst({
+            where: {
+              templateId: formID,
+              versionNumber,
+            },
+            select: {
+              id: true,
+              versionNumber: true,
+              status: true,
+              jsonConfig: true,
+            },
+          })
+          .catch((e) => prismaErrors(e, null))
+      : null;
+
+    if (versionNumber && !requestedVersion) return null;
+
     logEvent(user.id, { type: "Form", id: formID }, "ReadForm");
 
-    return parseTemplate(template);
-  } catch (e) {
+    return parseTemplate(template, {
+      version: requestedVersion ?? getBuilderVersion(template),
+    });
+  } catch {
     return null;
   }
 }
