+++ lib/templates/queries/getPublicTemplateByID.ts
@@ -1,11 +1,9 @@
 import { formCache } from "@lib/cache/formCache";
 import { prisma, prismaErrors } from "@gcforms/database";
 import { PublicFormRecord } from "@lib/types";
-import { mapTemplateToPublicFormRecord, parseTemplate } from "../internal";
+import { mapTemplateToPublicFormRecord, parseTemplate, templateRecordInclude } from "../internal";
 import { logMessage } from "@lib/logger";
-import { isTemplateVersioningEnabled } from "../versioning/internal";
-import { getPublicTemplateByID as getPublicTemplateByIDVersioningEnabled } from "@lib/templates/versioning/queries/getPublicTemplateByID";
-import { PublicTemplateMode } from "../versioning/internal/types";
+import type { PublicTemplateMode } from "../internal/types";
 
 /**
  * Get a form template by ID (only includes public information but does not require any permission)
@@ -16,17 +14,15 @@ export async function getPublicTemplateByID(
   formID: string,
   mode: PublicTemplateMode = "published"
 ): Promise<PublicFormRecord | null> {
-  if (await isTemplateVersioningEnabled()) {
-    return getPublicTemplateByIDVersioningEnabled(formID, mode);
-  }
-
   try {
     if (formCache.cacheAvailable) {
       // This value will always be the latest if it exists because
       // the cache is invalidated on change of a template
       const cachedValue = await formCache.check(formID);
       if (cachedValue) {
-        return cachedValue;
+        if (cachedValue.versionNumber !== undefined && cachedValue.versionNumber !== null) {
+          return cachedValue;
+        }
       }
     }
 
@@ -35,32 +31,21 @@ export async function getPublicTemplateByID(
         where: {
           id: formID,
         },
-        select: {
-          id: true,
-          created_at: true,
-          updated_at: true,
-          name: true,
-          jsonConfig: true,
-          isPublished: true,
-          deliveryOption: true,
-          securityAttribute: true,
-          formPurpose: true,
-          publishReason: true,
-          publishFormType: true,
-          publishDesc: true,
-          closingDate: true,
-          closedDetails: true,
-          saveAndResume: true,
-          ttl: true,
-          notificationsInterval: true,
-        },
+        include: templateRecordInclude,
       })
       .catch((e) => prismaErrors(e, null));
 
     // Short circuit the public record filtering if no form record is found or the form is marked as deleted (ttl != null)
     if (!template || template.ttl) return null;
 
-    const parsedTemplate = parseTemplate(template);
+    const selectedVersion =
+      mode === "published"
+        ? (template.currentPublishedVersion ?? null)
+        : (template.currentDraftVersion ?? template.currentPublishedVersion ?? null);
+
+    const parsedTemplate = parseTemplate(template, {
+      version: selectedVersion,
+    });
     const publicFormRecord = mapTemplateToPublicFormRecord(parsedTemplate);
 
     if (formCache.cacheAvailable) formCache.set(formID, publicFormRecord);
