--- lib/templates/versioning/queries/getTemplatePublishedStatus.ts
@@ -1,11 +1,0 @@
-import { getTemplateVersionState } from "./getTemplateVersionState";
-
-export async function getTemplatePublishedStatus(formID: string): Promise<boolean | null> {
-  const templateVersionState = await getTemplateVersionState(formID);
-
-  if (!templateVersionState) {
-    return null;
-  }
-
-  return templateVersionState.isPublished;
-}
